+++ pyjinhx/builtins/ui/pjx_card_header/pjx_card_header.py
@@ -2,6 +2,7 @@ from pyjinhx import BaseComponent
 from pyjinhx.base import AttrValue
 
 
-class PJXCard(BaseComponent):
+class PJXCardHeader(BaseComponent):
+    title: str = ""
     class_name: AttrValue = ""
     content: str | BaseComponent = ""